fix issue #11, segfault in shine_BF_freePartHolder
config->formatbits.side_info.headerPH and
config->formatbits.side_info.frameSIPH are allocated at the same time
thus checking one of them should be sufficient to ensure that an unsafe
free is not performed. Other solutions would be to check for NULL in
shine_BF_freePartHolder, but this is faster.